requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
numpy
groq
python-pptx